% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\name{find_studies}
\alias{find_studies}
\alias{get_studies}
\title{Extract available studies from the SQP 3.0 API}
\usage{
find_studies(study)

get_studies()
}
\arguments{
\item{study}{a string with the name of the study. Upper and lower cases are
ignored and regular expressions are supported.}
}
\value{
A two column \code{\link[tibble]{tibble}} with the id and name
of the study.
}
\description{
\code{find_studies} allows you to search for studies by names while
\code{get_studies} will return all available studies in the SQP 3.0 database.
}
\details{
The user should almost always start by using \code{find_studies}
which searches for a study based on it's name. If nothing useful comes up
then use \code{get_studies} which will return all available studies.

The returned \code{\link[tibble]{tibble}} will contain
the id and name of the study/studies. The user should take note of
the id of the desired study to checkout which questions are available
using \code{\link{get_questions}} and \code{\link{find_questions}}.
}
\examples{

\dontrun{

# Set your login information here. See ?sqp_login
sqp_login()

find_studies("ess")
find_studies("australia")

# or get_studies() for all studies

get_studies()

}
}
