% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{get_estimates}
\alias{get_estimates}
\alias{get_question_name}
\title{Extract variable estimates from the SQP prediction algorithm}
\usage{
get_estimates(id, all_columns = FALSE)

get_question_name(id)
}
\arguments{
\item{id}{a numeric vector containing the id(s) of variable(s) of interest. Can
be a one or more id's.}

\item{all_columns}{a logical stating whether to extract all available
columns from the SQP database. See the details section for a list of all possible variables.}
}
\value{
\code{get_estimates} returns \code{\link[tibble]{tibble}} with the predictions.
The number of columns depends on the \code{all_columns} argument.
\code{get_question_name} returns a character vector with the question name(s).
}
\description{
Extract variable estimates from the SQP prediction algorithm
}
\examples{

\dontrun{

# Log in with sqp_login first. See ?sqp_login

get_estimates(c(1, 2, 86))

get_estimates(c(1, 2, 86), all_columns = TRUE)

# Explore variable names

get_question_name(1)

get_question_name(1:10)

}

}
