% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqp_collect.R
\name{sqp_collect}
\alias{sqp_collect}
\title{Collect data from the Survey Quality Prediction database.}
\usage{
sqp_collect(...)
}
\arguments{
\item{...}{}
}
\description{
Collect data from the Survey Quality Prediction database.
}
