% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqp_correlate.R
\name{sqp_correlate}
\alias{sqp_correlate}
\title{Calculate a correlation matrix with an adjusted diagonal}
\usage{
sqp_correlate(x, diag_adj, use = "complete.obs", method = "pearson")
}
\arguments{
\item{x}{a matrix or data frame with numeric columns.}

\item{diag_adj}{a numeric vector with length equal to the number of columns of \code{x}}

\item{use}{an optional character string giving a method for computing covariances
in the presence of missing values. This must be (an abbreviation of) one of
the strings "everything", "all.obs", "complete.obs", "na.or.complete", or
"pairwise.complete.obs".}

\item{method}{a character string indicating which correlation coefficient
(or covariance) is to be computed. One of "pearson" (default), "kendall",
 or "spearman": can be abbreviated.}
}
\value{
a correlation \code{tibble} with variable names as a column and
the diagonal replaced by \code{diag_adj}
}
\description{
\code{sqp_correlate} calculates a correlation matrix through \code{\link{cor}}
and replaces the diagonal with the supplied numeric vector. It's a wrapper
around \code{\link{cor}} with slight tweaks.
}
\examples{

# New diagonal
new_diagonal <- rnorm(ncol(mtcars))

sqp_correlate(mtcars, new_diagonal)

sqp_correlate(mtcars, new_diagonal, method = "kendall")

}
\seealso{
\code{\link{cor}} for the workhorse behind the function and for
details on the \code{use} and \code{method} arguments.
}
