% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqpr-package.R
\docType{package}
\name{sqpr-package}
\alias{sqpr}
\alias{sqpr-package}
\title{sqpr: Obtain and analyze the quality of your survey questions with the SQP API}
\description{
Using the database of the SQP API, sqpr allows downloading data on the quality of over 40,000 questions, adjust for the Common Method Variance (CMV) of several variables, calculate the quality of the sumscore of several variables and adjust your modeling estimations for a variables measurement error.
}
\author{
\strong{Maintainer}: Jorge Cimentada \email{cimentadaj@gmail.com}

Authors:
\itemize{
  \item Willem Saris
}

Other contributors:
\itemize{
  \item Pompeu Fabra University [copyright holder, funder]
}

}
\keyword{internal}
