% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqp_cmv.R
\name{estimate_cmv}
\alias{estimate_cmv}
\title{Estimate the Common Method Variance (CMV) coefficient of a set of variables}
\usage{
estimate_cmv(sqp_data)
}
\arguments{
\item{sqp_data}{a data frame or tibble of class \code{sqp} given by \code{\link{sqp_collect}}
which contains the desired variables from which to estimate the CMV.}
}
\value{
a numeric vector of length one with the estimated coefficient
}
\description{
Estimate the Common Method Variance (CMV) coefficient of a set of variables
}
\examples{
library(tibble)

sqp_df <-
 tibble(question = paste0("V", 1:5),
 quality = c(0.2, 0.3, 0.5, 0.6, 0.9),
 reliability = c(0.2, 0.4, 0.5, 0.5, 0.7),
 validity = c(0.5, 0.1, 0.6, 0.7, 0.8))

estimate_cmv(sqp_df)

\dontrun{

sqp_df <-
 tibble(question = paste0("V", 1:5),
 quality = c(0.2, 0.3, 0.5, 0.6, 0.9),
 reliability = c(0.2, 0.4, 0.5, 0.5, 0.7),
 validity = c(NA, 0.1, 0.6, 0.7, 0.8))

estimate_cmv(sqp_df)
# Error in estimate_cmv(sqp_df) :
# `sqp_data` must have non-missing values at columns reliability and validity for all variables
}


}
\seealso{
\code{\link{sqp_cmv}} for automatically adjusting a correlation
matrix for the CMV and \code{\link{sqp_collect}} for obtaining SQP data.
}
