% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqp_rbind_metrics.R
\name{sqp_bind_metrics}
\alias{sqp_bind_metrics}
\title{Title}
\usage{
sqp_bind_metrics(df, question, metrics)
}
\arguments{
\item{metrics}{must be a named list containing only one numeric element per name.}
}
\description{
Title
}
