% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqp_cmv.R
\name{sqp_cmv}
\alias{sqp_cmv}
\title{Adjust a correlation matrix for Common Method Variance (CMV)}
\usage{
sqp_cmv(x, sqp_data, ..., standardized = TRUE, original_data, cmv = NULL)
}
\arguments{
\item{x}{a correlation matrix, a correlation data frame or a correlation
\code{tibble} given by \code{\link{sqp_correlate}}.}

\item{sqp_data}{a data frame or tibble of class \code{sqp} given by
\code{\link{get_estimates}}.}

\item{...}{two or more variables present in both \code{x} and \code{sqp_data}. Can
be both in bare unquoted names or as character strings.}

\item{standardized}{A logical stating whether to calculate a standardized Common Method Variance.
By default it's calculated as standardized. If \code{standardized} is set to \code{FALSE} then
the original CMV coefficient is multiplied by the standard deviation of the variables
in \code{original_data}. Whenever \code{standardized} is set to \code{FALSE}, \code{original_data}
must be a non-missing argument with a data frame that contains the variables specified in
\code{...}. If \code{standardized} is \code{TRUE} then \code{original_data} is ignored
even if supplied.}

\item{original_data}{A data frame containing the original data for variables specified
in \code{...}. For example, the original columns of the variables from the correlation
matrix supplied in \code{x}.}

\item{cmv}{an optional numeric vector of length 1 which contains the
CMV coefficient of the variables specified in \code{...}. It is strongly
suggested that this coefficient is estimated via \code{\link{estimate_cmv}}.
By default, it is set to NULL and it is calculated internally.}
}
\value{
the same matrix supplied in \code{x} but as a tibble with
the correlation coefficients of the variables supplied in \code{...}
adjusted for their common method variance.
If \code{x}  had row names, they're moved as a column called
'rowname'. If it doesnt have row names or a column named 'rowname' a new
column is created with the column names to mimic a correlation matrix.
}
\description{
\code{sqp_cmv} accepts a correlation matrix, a data frame matrix or a correlation
tibble from \code{\link{sqp_correlate}}
and adjusts the coefficients of the variables specified
in  \code{...} with the reliability and validity coefficients given
by \code{\link{get_estimates}}. All variables specified in \code{...} must
be present in both \code{x} and \code{sqp_data}. Optionally, you can supply
the cmv coefficient in the argument \code{cmv}.
}
\examples{

set.seed(2131)
library(tibble)

original_df <- as.data.frame(matrix(rnorm(100, sd = 50), nrow = 20))
corr_tibble <- sqp_correlate(original_df, rnorm(5))

# Toy dataset
sqp_df <-
 tibble(question = paste0("V", 1:5),
 quality = c(0.2, 0.3, 0.5, 0.6, 0.9),
 reliability = c(NA, 0.4, 0.5, 0.5, 0.7),
 validity = c(NA, NA, 0.6, 0.7, 0.8))

sqp_df <- structure(sqp_df, class = c(class(sqp_df), "sqp"))

# Show that when y is not from sqp class, there's an error

# Original correlation matrix with new diagonal
corr_tibble

# Coefficient of correlation changes
# when adjusting for common method variance
sqp_cmv(corr_tibble, sqp_df, V4, V5)

# The V5*V4 from both the upper/lower triangles
# correlation matrix changed from -0.05 to -0.203

# If you want to get a standardized CMV, then supply
# `standardized = FALSE` and `original_data`.

sqp_cmv(corr_tibble, sqp_df, V4, V5,
        standardized = FALSE,
        original_data = original_df)

}
\seealso{
\code{\link{sqp_correlate}} for calculating correlation matrices,
\code{\link{get_estimates}} for obtaining the SQP estimates directly into R
and \code{\link{estimate_cmv}} for calculating the CMV manually.
}
