% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqp_collect.R
\name{sqp_collect}
\alias{sqp_collect}
\title{Collect data from the Survey Quality Prediction database.}
\usage{
sqp_collect(...)
}
\arguments{
\item{...}{variable names to collect from the SQP database}
}
\value{
a tibble with the selected variables as rows and SQP metrics
as columns
}
\description{
Collect data from the Survey Quality Prediction database.
}
\examples{

5 + 5

}
