% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqpr-package.R
\docType{package}
\name{sqpr-package}
\alias{sqpr}
\alias{sqpr-package}
\title{sqpr: Extract the Quality of your Survey Questions with SQP 3.0}
\description{
Using the database of the SQP 3.0 API, sqpr allows downloading data on the quality of over 40,000 questions, including the quality, validity and reliability of each question.
}
\author{
\strong{Maintainer}: Jorge Cimentada \email{cimentadaj@gmail.com}

Authors:
\itemize{
  \item Wiebke Weber
}

}
\keyword{internal}
